requirements.txt
openai>=1.0.0
anthropic>=0.39.0
python-dotenv>=1.0.0
pytest>=7.0.0
numpy>=2.0.0
rdkit>=2024.3.5
pubchempy>=1.0.4
